+++ zcash_primitives/CHANGELOG.md
@@ -6,6 +6,15 @@ and this library adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `zcash_primitives::transaction`:
+  - `Transaction::temporary_zcashd_read_v5_sapling`
+  - `Transaction::temporary_zcashd_write_v5_sapling`
+- Implementations of `memuse::DynamicUsage` for the following types:
+  - `zcash_primitives::transaction::components::sapling`:
+    - `Bundle<Authorized>`
+    - `SpendDescription<Authorized>`
+
 ### Changed
 - MSRV is now 1.65.0.
 - Bumped dependencies to `secp256k1 0.26`, `hdwallet 0.4`.